requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.6.0
boolean.py==5.0
CacheControl==0.14.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
cyclonedx-python-lib==9.1.0
defusedxml==0.7.1
fastapi==0.115.14
filelock==3.18.0
flake8==7.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
license-expression==30.4.3
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
msgpack==1.1.1
packageurl-python==0.17.1
packaging==25.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.3.8
pluggy==1.6.0
py-serializable==2.0.0
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
ruff==0.12.1
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.46.2
toml==0.10.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
